put .java files in respective packages
