+++ src/main/java/duke/commands/MarkTaskNotDone.java
@@ -1,0 +1,16 @@
+package duke.commands;
+
+import duke.exceptions.InvalidIndexException;
+import duke.exceptions.StorageException;
+import duke.mainUtils.Storage;
+import duke.mainUtils.Ui;
+import duke.tasks.TaskList;
+
+public class MarkTaskNotDone extends Command {
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) throws InvalidIndexException, StorageException {
+        String[] userInput = ui.getCommand();
+        int index = Integer.parseInt(userInput[userInput.length - 1]) - 1;
+        taskList.getTask(index).markNotDone();
+    }
+}
