+++ src/main/java/duke/exceptions/DukeException.java
@@ -1,0 +1,8 @@
+package duke.exceptions;
+
+public class DukeException extends Exception{
+    public DukeException(String message) {
+        super("Error: " + message);
+    }
+
+}
