+++ src/main/java/duke/exceptions/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package duke.exceptions;
+
+public class InvalidCommandException extends DukeException {
+    public InvalidCommandException() {
+        super("Invalid command. Use help to for more info");
+    }
+}
