+++ src/main/java/duke/exceptions/InvalidDateException.java
@@ -1,0 +1,7 @@
+package duke.exceptions;
+
+public class InvalidDateException extends DukeException {
+    public InvalidDateException() {
+        super("Invalid date format. Please use dd-MM-yyyy");
+    }
+}
