+++ src/main/java/duke/exceptions/InvalidTaskException.java
@@ -1,0 +1,7 @@
+package duke.exceptions;
+
+public class InvalidTaskException extends DukeException {
+    public InvalidTaskException() {
+        super("Invalid task. Use help for guidance.");
+    }
+}
