+++ src/main/java/duke/exceptions/StorageException.java
@@ -1,0 +1,7 @@
+package duke.exceptions;
+
+public class StorageException extends DukeException {
+    public StorageException() {
+        super("Error loading/saving file.");
+    }
+}
