+++ src/main/java/duke/fileUtils/FilePaths.java
@@ -1,0 +1,16 @@
+package duke.fileUtils;
+
+public final class FilePaths {
+
+    public static final String LOGO_PATH = "./src/duke.main/graphics/logo.txt";
+
+    public static final String HELP_PATH = "./src/duke.main/graphics/help.txt";
+
+    public static final String HORIZONTAL_LINE_PATH = "./src/duke.main/graphics/horizontal_line.txt";
+
+    public static final String TASKS_SAVE_FILE_PATH = "./src/duke.main/data/duke.tasks.txt";
+
+    public static final String ERROR_GRAPHIC_PATH = "./src/duke.main/graphics/error_graphic.txt";
+
+
+}
