+++ src/main/java/duke/tasks/DeadlineTask.java
@@ -1,0 +1,42 @@
+package duke.tasks;
+
+import java.time.LocalDate;
+
+public class DeadlineTask extends Task {
+    protected LocalDate doneBy;
+
+    public DeadlineTask(String description, LocalDate doneBy) {
+        super(description);
+        this.doneBy = doneBy;
+    }
+
+    @Override
+    public void updateDoneIcon() {
+        super.updateDoneIcon();
+    }
+
+    @Override
+    public String getDoneIcon() {
+        return super.getDoneIcon();
+    }
+
+    public LocalDate getDoneBy() {
+        return this.doneBy;
+    }
+
+    @Override
+    public void markDone() {
+        super.markDone();
+    }
+
+    @Override
+    public void markNotDone() {
+        super.markNotDone();
+    }
+
+    @Override
+    public String toString() {
+        return "D" + " | " + super.toString() + " | " + "(by: " + getDoneBy() + ")";
+    }
+
+}
