--- src/main/java/exceptions/DukeException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class DukeException extends Exception{
-    public DukeException(String message) {
-        super("Error: " + message);
-    }
-
-}
