--- src/main/java/exceptions/InvalidCommandException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidCommandException extends DukeException{
-    public InvalidCommandException() {
-        super("Invalid command. Use help to for more info");
-    }
-}
