--- src/main/java/exceptions/InvalidDateException.java
@@ -1,6 +1,0 @@
-package exceptions;
-public class InvalidDateException extends DukeException{
-    public InvalidDateException() {
-        super("Invalid date format. Please use dd-MM-yyyy");
-    }
-}
