--- src/main/java/exceptions/InvalidIndexException.java
@@ -1,6 +1,0 @@
-package exceptions;
-public class InvalidIndexException extends DukeException{
-    public InvalidIndexException() {
-        super("Index out of bounds. Use list to view all tasks.");
-    }
-}
