--- src/main/java/exceptions/InvalidTaskException.java
@@ -1,6 +1,0 @@
-package exceptions;
-public class InvalidTaskException extends DukeException{
-    public InvalidTaskException() {
-        super("Invalid task. Use help for guidance.");
-    }
-}
