--- src/main/java/exceptions/StorageException.java
@@ -1,7 +1,0 @@
-package exceptions;
-public class StorageException extends DukeException{
-    public StorageException() {
-        super("Error loading/saving file.");
-        printStackTrace();
-    }
-}
