--- src/main/java/fileUtils/FilePaths.java
@@ -1,16 +1,0 @@
-package fileUtils;
-
-public final class FilePaths {
-
-    public static final String LOGO_PATH = "./src/main/graphics/logo.txt";
-
-    public static final String HELP_PATH = "./src/main/graphics/help.txt";
-
-    public static final String HORIZONTAL_LINE_PATH = "./src/main/graphics/horizontal_line.txt";
-
-    public static final String TASKS_SAVE_FILE_PATH = "./src/main/data/tasks.txt";
-
-    public static final String ERROR_GRAPHIC_PATH = "./src/main/graphics/error_graphic.txt";
-
-
-}
